fix schedule

+++ database/migrations/2024_10_23_010011_add_total_order_field_on_events_table.php
@@ -9,7 +9,7 @@ class AddTotalOrderFieldOnEventsTable extends Migration
     public function up()
     {
         Schema::table('events', function (Blueprint $table) {
-            $table->decimal('total_order')->nullable()->after('status');
+            $table->integer('total_order')->nullable()->after('status');
         });
     }
